Get cover dimension from active document
This is great if you need to create a paperback cover for a set of
pages  or if you want to rebuild a case that has been build using the
gutter to indicate spine. Which was the preferred way of building
pre-cs5. This is still a work in progress and need to be finished. Use
with caution!